pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "money-flow-service"
version = "0.1.0"
description = "Excel-driven household finance web service rebuilt from scratch."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "email-validator>=2.2.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "openpyxl>=3.1.5",
  "pydantic>=2.12.0",
  "pydantic-settings>=2.11.0",
  "PyJWT>=2.10.0",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.44",
  "psycopg[binary]>=3.2.0",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "pytest-asyncio>=1.2.0",
  "ruff>=0.14.0",
]

[tool.pytest.ini_options]
pythonpath = ["backend"]
testpaths = ["backend/tests"]
addopts = "-q"

[tool.setuptools.packages.find]
where = ["backend"]
include = ["app*"]
